+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 Changelog
 ---------
 
-# Unreleased
+# 10.5.4
 - Use `process.nextTick` instead of `setTimeout` for deferring calls to intersection observer callback
 - Never run observer callback on unobserved elements